requirements-pyqt5.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
easydict==1.9
gast==0.3.3
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.35.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.4.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.2
Markdown==3.3.3
matplotlib==3.3.4
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==3.3.0
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3.0.2
pytesseract==0.3.7
python-dateutil==2.8.1
python-dotenv==0.15.0
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scipy==1.4.1
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.3.0rc0
termcolor==1.1.0
tf-estimator-nightly==2.3.0.dev2020062301
tqdm==4.56.2
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
